refactor: rename name files to standard format (eg. cat.ts to Cat.ts) and update import paths for consistency
