+++ lib/interfaces/Cat.ts
