+++ lib/interfaces/CatDetails.ts
