+++ lib/interfaces/CatStats.ts
