+++ lib/interfaces/HealthLog.ts
