+++ lib/interfaces/PastReport.ts
