+++ lib/interfaces/PerCatNotificationPref.ts
