+++ lib/interfaces/ReportConfig.ts
