+++ lib/interfaces/Session.ts
